fix : Id값 형 수정(Integer -> Long)

+++ src/main/java/nextstep/subway/dto/LineRequest.java
@@ -7,19 +7,19 @@ public class LineRequest {
 
     private String name;
     private String color;
-    private Integer upStationId;
-    private Integer downStationId;
+    private Long upStationId;
+    private Long downStationId;
     private Integer distance;
 
     public Line toLine(Station upStation, Station downStation) {
         return Line.of(name, color, upStation, downStation, distance);
     }
 
-    public Integer getUpStationId() {
+    public Long getUpStationId() {
         return upStationId;
     }
 
-    public Integer getDownStationId() {
+    public Long getDownStationId() {
         return downStationId;
     }
 }
